+++ src/pages/loginPage/Login.module.css
@@ -1,50 +1,50 @@
-.cont_login {
+.main_login {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    background-color: #0F0B19;
+    border-bottom-left-radius: 50px;
+    border-bottom-right-radius: 50px;
+    height: 90vh;
+}
+
+.logo_login {
+    width: 100vw;
     text-align: center;
-    width: 390px;
-    margin: 0 auto;
+    padding-top: 180px;
+    flex-grow: 1;
 }
 
-.tit_login {
-    padding-top: 60px;
-    font-size: 24px;
+.cont_login {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-grow: 1;
 }
 
-.label_login {
-    display: block;
-    color: #767676;
-    text-align: left;
-    width: 322px;
-    margin: 0 auto;
+.ul_login {
+    display: flex;
 }
 
-.form_login {
-    margin-top: 50px;
-}
-
-.input_login {
-    width: 322px;
-    border-left-width: 0;
-    border-right-width: 0;
-    border-top-width: 0;
-    border-bottom-width: 1px;
-    margin: 10px 0 30px 0;
+.email {
+    color: red;
 }
 
 .btn_login {
-    display: block;
-    margin: 10px auto 20px;
-    width: 322px;
-    border: 1px solid #0f0b19;
-    border-radius: 25px;
-    padding: 10px 0px;
+    color: #c4c4c4;
     cursor: pointer;
 }
 
+.btn_login::after {
+    display: inline-block;
+    margin-left: 12px;
+    content: "";
+    width: 1px;
+    height: 10px;
+    background-color: #c4c4c4;
+}
+
 .btn_join {
-    display: block;
-    margin: 20px auto;
-    text-align: center;
-    font-size: 14px;
-    color: #767676;
+    color: #c4c4c4;
     cursor: pointer;
 }
