+++ src/pages/loginPage/loginEmail.module.css
@@ -1,0 +1,50 @@
+.cont_login {
+    text-align: center;
+    width: 390px;
+    margin: 0 auto;
+}
+
+.tit_login {
+    padding-top: 60px;
+    font-size: 24px;
+}
+
+.label_login {
+    display: block;
+    color: #767676;
+    text-align: left;
+    width: 322px;
+    margin: 0 auto;
+}
+
+.form_login {
+    margin-top: 50px;
+}
+
+.input_login {
+    width: 322px;
+    border-left-width: 0;
+    border-right-width: 0;
+    border-top-width: 0;
+    border-bottom-width: 1px;
+    margin: 10px 0 30px 0;
+}
+
+.btn_login {
+    display: block;
+    margin: 10px auto 20px;
+    width: 322px;
+    border: 1px solid #0f0b19;
+    border-radius: 25px;
+    padding: 10px 0px;
+    cursor: pointer;
+}
+
+.btn_join {
+    display: block;
+    margin: 20px auto;
+    text-align: center;
+    font-size: 14px;
+    color: #767676;
+    cursor: pointer;
+}
